# storage/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is swagger documentation for store store",
    "version": "0.0.1",
    "title": "Store Backend",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "miqdad.farcha@store.backend.fdc"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "host": "store.backend.fdc",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/product-category/list-by-merchant": {
      "get": {
        "tags": [
          "product-category"
        ],
        "summary": "Function to get list of product category by Merchant.",
        "description": "",
        "operationId": "getAllProductCategoryByMerchant",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "merchant_id",
            "in": "query",
            "description": "The merchant ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
    "/product-category/position": {
      "get": {
        "tags": [
          "product-category"
        ],
        "summary": "Function to get product categories posiition",
        "description": "",
        "operationId": "getCategoryPosition",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
    "/product-categorys/position": {
      "post": {
        "tags": [
          "product-category"
        ],
        "summary": "update product category position",
        "description": "",
        "operationId": "updateCategoryPosition",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "formData",
            "description": "The category ID.",
            "required": true,
            "type": "integer"
          },
          {
            "name": "position",
            "in": "formData",
            "description": "The category position.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              },
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              }
            },
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          },
          "401": {
            "description": "Unauthorized access"
          }
        }
      }
    },
...
  },
...
}